Add some automation to wheel builds and relocation.

+++ setup.py
@@ -1,4 +1,14 @@
 #!/usr/bin/env python
+
+# Two environmental variables influence this script.
+#
+# GDAL_CONFIG: the path to a gdal-config program that points to GDAL headers,
+# libraries, and data files.
+#
+# PACKAGE_DATA: if defined, GDAL and PROJ4 data files will be copied into the
+# source or binary distribution. This is essential when creating self-contained
+# binary wheels.
+
 import logging
 import os
 import shutil
@@ -64,16 +74,29 @@ try:
         else:
             # e.g. -framework GDAL
             extra_link_args.append(item)
-    shutil.rmtree('rasterio/gdal_data')
-    shutil.copytree(datadir, 'rasterio/gdal_data')
+
+    # Conditionally copy the GDAL data. To be used in conjunction with
+    # the bdist_wheel command to make self-contained binary wheels.
+    if os.environ.get('PACKAGE_DATA'):
+        try:
+            shutil.rmtree('rasterio/gdal_data')
+        except OSError:
+            pass
+        shutil.copytree(datadir, 'rasterio/gdal_data')
+
 except Exception as e:
     log.warning("Failed to get options via gdal-config: %s", str(e))
 
-# Copy PROJ.4 data.
-projdatadir = '/usr/local/share/proj'
-if os.path.exists(projdatadir):
-    shutil.rmtree('rasterio/proj_data')
-    shutil.copytree(projdatadir, 'rasterio/proj_data')
+# Conditionally copy PROJ.4 data. Presumes PROJ.4 is installed locally
+# with --prefix=/usr/local.
+if os.environ.get('PACKAGE_DATA'):
+    projdatadir = '/usr/local/share/proj'
+    if os.path.exists(projdatadir):
+        try:
+            shutil.rmtree('rasterio/proj_data')
+        except OSError:
+            pass
+        shutil.copytree(projdatadir, 'rasterio/proj_data')
 
 ext_options = dict(
     include_dirs=include_dirs,
@@ -141,38 +164,41 @@ inst_reqs = [
 if sys.version_info < (3, 4):
     inst_reqs.append('enum34')
 
-setup(name='rasterio',
-      version=version,
-      description=(
-          "Fast and direct raster I/O for Python programmers who use Numpy"),
-      long_description=readme,
-      classifiers=[
-          'Development Status :: 4 - Beta',
-          'Intended Audience :: Developers',
-          'Intended Audience :: Information Technology',
-          'Intended Audience :: Science/Research',
-          'License :: OSI Approved :: BSD License',
-          'Programming Language :: C',
-          'Programming Language :: Python :: 2.6',
-          'Programming Language :: Python :: 2.7',
-          'Programming Language :: Python :: 3.3',
-          'Programming Language :: Python :: 3.4',
-          'Topic :: Multimedia :: Graphics :: Graphics Conversion',
-          'Topic :: Scientific/Engineering :: GIS'
-          ],
-      keywords='raster gdal',
-      author='Sean Gillies',
-      author_email='sean@mapbox.com',
-      url='https://github.com/mapbox/rasterio',
-      license='BSD',
-      package_dir={'': '.'},
-      packages=['rasterio', 'rasterio.rio'],
-      entry_points='''
+setup_args = dict(
+    name='rasterio',
+    version=version,
+    description="Fast and direct raster I/O for use with Numpy and SciPy",
+    long_description=readme,
+    classifiers=[
+        'Development Status :: 4 - Beta',
+        'Intended Audience :: Developers',
+        'Intended Audience :: Information Technology',
+        'Intended Audience :: Science/Research',
+        'License :: OSI Approved :: BSD License',
+        'Programming Language :: C',
+        'Programming Language :: Python :: 2.6',
+        'Programming Language :: Python :: 2.7',
+        'Programming Language :: Python :: 3.3',
+        'Programming Language :: Python :: 3.4',
+        'Topic :: Multimedia :: Graphics :: Graphics Conversion',
+        'Topic :: Scientific/Engineering :: GIS'],
+    keywords='raster gdal',
+    author='Sean Gillies',
+    author_email='sean@mapbox.com',
+    url='https://github.com/mapbox/rasterio',
+    license='BSD',
+    package_dir={'': '.'},
+    packages=['rasterio', 'rasterio.rio'],
+    entry_points='''
         [console_scripts]
         rio=rasterio.rio.main:cli
-      ''',
-      include_package_data=True,
-      package_data={'rasterio': ['gdal_data/*', 'proj_data/*']},
-      ext_modules=ext_modules,
-      zip_safe=False,
-      install_requires=inst_reqs )
+    ''',
+    include_package_data=True,
+    ext_modules=ext_modules,
+    zip_safe=False,
+    install_requires=inst_reqs)
+
+if os.environ.get('PACKAGE_DATA'):
+    setup_args['package_data'] = {'rasterio': ['gdal_data/*', 'proj_data/*']}
+
+setup(**setup_args)
